requirements.txt
pip==21.1.2
setuptools==57.0.0
wheel==0.36.2
appdirs==1.4.3
Django==3.2.10
asgiref==3.4.1
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2